refactor(rpc): move Serialize into file
